# Introduce overrides of protocol members and drop them from witness tables