--- a/include/swift/AST/Attr.def
+++ b/include/swift/AST/Attr.def
@@ -375,6 +375,10 @@ SIMPLE_DECL_ATTR(_hasInitialValue, HasInitialValue,
   OnVar |
   UserInaccessible,
   78)
+SIMPLE_DECL_ATTR(_nonoverride, NonOverride,
+  OnFunc | OnAccessor | OnVar | OnSubscript | OnConstructor | OnAssociatedType |
+  NotSerialized,
+  79)
 
 #undef TYPE_ATTR
 #undef DECL_ATTR_ALIAS