--- a/include/swift/AST/Attr.def
+++ b/include/swift/AST/Attr.def
@@ -377,7 +377,7 @@ SIMPLE_DECL_ATTR(_hasInitialValue, HasInitialValue,
   78)
 SIMPLE_DECL_ATTR(_nonoverride, NonOverride,
   OnFunc | OnAccessor | OnVar | OnSubscript | OnConstructor | OnAssociatedType |
-  NotSerialized,
+  UserInaccessible | NotSerialized,
   79)
 
 #undef TYPE_ATTR