… lib/Sema/TypeChecker.h
… lib/Serialization/Deserialization.cpp
… stdlib/public/core/ArrayBufferProtocol.swift
… stdlib/public/core/ArrayType.swift
… stdlib/public/core/BidirectionalCollection.swift
… stdlib/public/core/Collection.swift
… stdlib/public/core/FloatingPoint.swift.gyb
… stdlib/public/core/Integers.swift
… stdlib/public/core/MutableCollection.swift
… stdlib/public/core/RandomAccessCollection.swift
… stdlib/public/core/RangeReplaceableCollection.swift
… stdlib/public/core/ShadowProtocols.swift
… test/Compatibility/attr_override.swift
… test/IDE/complete_associated_types.swift
… test/Index/conformances.swift
… test/Index/roles.swift
… test/SILGen/witness_table_overrides.swift
… test/SILOptimizer/sil_combine.sil
… test/attr/attr_override.swift
… test/decl/class/nonoverride.swift
… test/decl/protocol/override.swift
… test/decl/protocol/warn_override.swift