When a protocol restates a requirement from its inherited protocol, track that as an override the same way we do with overrides in classes. Such overrides are recorded in the AST but are omitted from the witness table itself. This makes restating protocol requirements in inheriting protocols ABI-neutral, so we can restate protocol requirements to influence associated type inference without baking those decisions into the ABI.

As part of this, introduce tools that allow us to audit all of the restated protocol requirements within the standard library, via a warning new warning flag (`-warn-implicit-overrides`) to warn about such requirements, as well as an appropriate suppression mechanism (`override` to treat it as an override, `@_nonoverride` to give it a new witness table entry). Only those restated requirements that are meant to have a semantic effect, like `BidirectionalCollection.index(_:offsetBy:)`, should use `@_nonoverride`, because it commits us to having that entry in the witness table forever.

Rolling this out in the standard library shrinks the size of the standard library binary by 196k.

Implements rdar://problem/43870489.