@DougGregor 2018-08-31
@swift-ci please test source compatibility