@DougGregor 2018-09-04
@airspeedswift  and @moiseev , https://github.com/apple/swift/pull/19034/commits/7fa0b649912b44e8a21c9bcccd876344b7e6d6e6 audits all of the "overriding" protocol requirements to decide which have semantic effect (like `BidirectionalCollection.index(_:offsetBy:)`, marked with `@_nonoverride`) vs. those that are there only to paper over bugs in associated type inference (like `BidirectionalCollection.startIndex`, marked with `override`). I would appreciate a review to verify that I got the set right.