@DougGregor 2018-09-04
@swift-ci please benchmark