@DougGregor 2018-09-05
@swift-ci please smoke test