@DougGregor 2018-09-05
@swift-ci please test source compatibility