@DougGregor 2018-09-05
@airspeedswift the `IterateData` result is interesting. Maybe there's `@_nonoverride` we could add somewhere to improve performance, but I'm not sure if it matters that much.