@airspeedswift 2018-09-05
The `IterateDate` benchmark is noisy, it's one of the few that still sometimes crops up on unrelated benchmarks.