@DougGregor 2018-09-05
Okay. Assuming I didn't break something w.r.t. the compatibility suite or compile-time performance, I'll go ahead and land this.