@DougGregor 2018-09-05
The source-compatibility failure is covered by https://github.com/apple/swift/pull/19146